[core] Split  into  and
